pyproject.toml
[project]
name = "diplomacy-grpo-pipeline"
version = "0.1.0"
description = "Self-play pipeline for Diplomacy using GRPO with country-based grouping"
authors = [
    {name = "Oscar", email = "oscar@example.com"},
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = {text = "MIT"}

dependencies = [
    # Core dependencies
    "torch>=2.6.0", 
    "transformers>=4.36.0",
    "datasets>=3.6.0",
    "accelerate>=0.28.0",
    "peft>=0.8.0",
    
    # GRPO and RL training
    "trl>=0.17.0",
    "wandb>=0.16.0",
    
    # Async and networking
    "openai>=1.81.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Utilities
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

vllm = [
    "vllm>=0.8.5",
    "flash-attn>=2.0.0",
]

all = [
    "diplomacy-grpo-pipeline[dev,vllm]",
    "deepspeed>=0.12.0",
    "liger-kernel>=0.5.10",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["diplomacy_grpo", "diplomacy_grpo.core", "diplomacy_grpo.integration", "diplomacy_grpo.rewards"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--verbose",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow tests (>1s)",
    "gpu: Tests requiring GPU",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Exclude external repositories
  AI_Diplomacy
  | verifiers
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    "AI_Diplomacy",
    "verifiers", 
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "^AI_Diplomacy/",
    "^verifiers/",
]